# internal/central/admin/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChargePilot Go admin API",
    "version": "2026-09-30",
    "description": "已接线的 Go 后台认证、仪表盘和现有 PC 页面接口。字段与工作流见 docs/api/go-admin-pages.md；不代表目标设计中的全部后台能力。"
  },
  "paths": {
...
    "/api/v1/admin/stations": {
      "get": {
        "summary": "GET /api/v1/admin/stations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "station.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        }
      },
      "post": {
        "summary": "POST /api/v1/admin/stations",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "station.create",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        }
      }
    },
    "/api/v1/admin/stations/{id}": {
      "get": {
        "summary": "GET /api/v1/admin/stations/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "station.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "PUT /api/v1/admin/stations/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "station.update",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/admin/devices": {
      "get": {
        "summary": "GET /api/v1/admin/devices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "device.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "station_id",
            "in": "query",
            "required": false,
            "description": "固定站点范围；必须为单个正整数，空值和重复参数返回 400。",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "description": "按 keyword/status/station_id 分页读取设备；同时应用当前账号站点与厂商范围。"
      }
    },
    "/api/v1/admin/devices/{id}": {
      "get": {
        "summary": "GET /api/v1/admin/devices/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "device.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "summary": "GET /api/v1/admin/orders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "order.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        }
      }
    },
    "/api/v1/admin/orders/{id}": {
      "get": {
        "summary": "GET /api/v1/admin/orders/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "order.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}/timeline": {
      "get": {
        "summary": "GET /api/v1/admin/orders/{id}/timeline",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "order.read",
        "responses": {
          "200": {
            "description": "成功，统一 code/message/data/request_id/trace_id 响应"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "未登录或会话失效"
          },
          "403": {
            "description": "当前实时权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "业务状态或幂等冲突"
          },
          "503": {
            "description": "依赖服务不可用"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/stations/{id}/split-template": {
      "put": {
        "summary": "绑定站点分账模板",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission": "station.update + finance.split_template.create",
        "description": "站点须停用；已有交易历史的设备站点不可改绑。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "template_id",
                  "expected_template_id"
                ],
                "properties": {
                  "template_id": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "expected_template_id": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "参数无效"
          },
          "401": {
            "description": "会话失效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "模板已绑定或并发条件冲突"
          },
          "503": {
            "description": "存储不可用"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}